feat: improve filter options and add index filter
Adds a filter to the index page to allow users to
filter down the list of benchmarks.

Also updates the benchmark page filter to only allow
the user to select filters that result in actual data.
This fixes a bug where for configs that are sparse,
sometimes the user would select a filter that resulted in no data.